change tg lib

+++ internal/model/model.go
@@ -1,13 +1,13 @@
 package model
 
 type Message struct {
-	CreatedAt int
+	CreatedAt int64
 	MessageId string
 	Text      string
 }
 
 type Transaction struct {
-	CreatedAt int
+	CreatedAt int64
 	MessageId string
 	Amount    float64
 	Category  string
